Cambios + sprites

+++ Juego/JuegoTDP19/src/Entidad/EnemigoGolemHielo.java
@@ -8,7 +8,7 @@ import Inteligencia.Inteligencia;
 
 public class EnemigoGolemHielo extends Enemigo {
 
-	public EnemigoGolemHielo(Point pos, int vida, int damage, float velocidadAtaque, int velocidad, int puntos) {
+	public EnemigoGolemHielo(Point pos, int vida, int damage, float velocidadAtaque, double velocidad, int puntos) {
 
 		super(pos, vida, damage, velocidadAtaque, velocidad, puntos);
 		imagen[0] = new ImageIcon("Sprites\\Golems\\GolemHieloGIF\\aaa4.gif");
